chore(Exception): checkNumeric, checkDivisible (src/main/java/lotto/Exception.java)
checkNumeric과 checkDivisible 메소드 수정:
-돈의 범위이기 때문에 큰 수가 사용될 것 같아 변수의 범위를 int 에서 long으로 바꿈

+++ src/main/java/lotto/Exception.java
@@ -1,10 +1,10 @@
 package lotto;
 
 abstract public class Exception extends Throwable {
-    public int checkNumeric(String input){
-        int number;
+    public long checkNumeric(String input){
+        long number;
         try{
-            number = Integer.parseInt(input);
+            number = Long.parseLong(input);
         }
         catch (NumberFormatException ex) {
             throw new IllegalArgumentException();
@@ -12,8 +12,8 @@ abstract public class Exception extends Throwable {
         return number;
     }
 
-    public void checkDivisible(int dividend, int divisor){
-        int remainder = dividend % divisor;
+    public void checkDivisible(long dividend, long divisor){
+        long remainder = dividend % divisor;
         if(remainder!=0) throw new IllegalArgumentException();
     }
 }
